refactor:parse command
