+++ Makefile
@@ -32,7 +32,7 @@ SRCS = srcs/main.c \
        srcs/parser/lexer.c \
        srcs/parser/lexer_utils.c \
        srcs/parser/parser.c \
-       srcs/parser/parser_utils.c \
+       srcs/parser/parse_command.c \
        srcs/parser/parse_redirection.c \
        srcs/parser/tokens.c \
        srcs/parser/ast.c \
